Fixed failing test cases and refactored Compound's standalone stories

+++ contracts/protocol/defiAdapters/compound/CompoundAdapter.sol
@@ -250,7 +250,7 @@ contract CompoundAdapter is IAdapter, IAdapterHarvestReward, IAdapterInvestLimit
                 )
             );
         } else {
-            _codes[2] = abi.encode(_liquidityPool, abi.encodeWithSignature("mint(uint256)", uint256(_depositAmount)));
+            _codes[2] = abi.encode(_liquidityPool, abi.encodeWithSignature("mint(uint256)", _depositAmount));
         }
     }
